Move core functionality to index.js
WIP

refactor: move core functionality to index.js

suppressing output for CI

adding back in missing return statement

removing stray console.log

putting output back, sorry CI

rename IPFS log to OrbitDB

Removing rimraf

moving files to src

moving report and benchmarks folder

Adding global gc to index.js

linting
