+++ package.json
@@ -2,12 +2,12 @@
   "name": "orbit-db-benchmark-runner",
   "version": "1.0.4",
   "description": "OrbitDB Benchmark Runner",
-  "main": "index.js",
+  "main": "./src/index.js",
   "bin": {
-    "benchmark-runner": "./cli.js"
+    "benchmark-runner": "./src/cli.js"
   },
   "scripts": {
-    "test": "nyc mocha --expose-gc"
+    "test": "nyc mocha"
   },
   "author": "mistakia",
   "license": "MIT",
