+++ test/benchmarks/index.js
@@ -1,0 +1,5 @@
+const consoleLog = require('./console-log')
+
+module.exports = [
+  ...consoleLog
+]
