+++ test/cli.test.js
@@ -1,8 +1,11 @@
 /* eslint-disable no-unused-expressions */
 
 const yargs = require('yargs')
-const cli = require('../cli')
+const cli = require('../src/cli')
 const expect = require('chai').expect
+const { start } = require('../src/index')
+
+const benchmarks = require(process.cwd() + '/test/benchmarks')
 
 describe('CLI Test', () => {
   it('should return help output', async () => {
@@ -24,7 +27,7 @@ describe('CLI Test', () => {
   })
 
   describe('baseline', () => {
-    it('should parse basline value', () => {
+    it('should parse baseline value', () => {
       const yargsCmd = yargs.command(cli)
       const yargsResult = yargsCmd.parse(
         'cli --baseline', {}
@@ -33,7 +36,7 @@ describe('CLI Test', () => {
       expect(yargsResult.b).to.be.true
     })
 
-    it('should parse basline value when using alias', () => {
+    it('should parse baseline value when using alias', () => {
       const yargsCmd = yargs.command(cli)
       const yargsResult = yargsCmd.parse(
         'cli -b', {}
@@ -41,9 +44,20 @@ describe('CLI Test', () => {
       expect(yargsResult.baseline).to.be.true
       expect(yargsResult.b).to.be.true
     })
+
+    it('should run a baseline test', async () => {
+      const yargsCmd = yargs.command(cli)
+      const yargsResult = yargsCmd.parse(
+        'cli --baseline', {}
+      )
+
+      await start(benchmarks, yargsResult)
+    })
   })
 
-  describe('report', () => {
+  describe('report', function () {
+    this.timeout(6000)
+
     it('should parse report value', () => {
       const yargsCmd = yargs.command(cli)
       const yargsResult = yargsCmd.parse(
@@ -61,9 +75,18 @@ describe('CLI Test', () => {
       expect(yargsResult.report).to.be.true
       expect(yargsResult.r).to.be.true
     })
+
+    it('should generate report', async () => {
+      const yargsCmd = yargs.command(cli)
+      const yargsResult = yargsCmd.parse(
+        'cli -r', {}
+      )
+
+      await start(benchmarks, yargsResult)
+    })
   })
 
-  describe('list', () => {
+  describe('list', function () {
     it('should parse list value', () => {
       const yargsCmd = yargs.command(cli)
       const yargsResult = yargsCmd.parse(
@@ -80,6 +103,15 @@ describe('CLI Test', () => {
       )
       expect(yargsResult.list).to.be.true
       expect(yargsResult.l).to.be.true
+    })
+
+    it('should list benchmarks', async () => {
+      const yargsCmd = yargs.command(cli)
+      const yargsResult = yargsCmd.parse(
+        'cli -l', {}
+      )
+
+      await start(benchmarks, yargsResult)
     })
   })
 
@@ -105,9 +137,7 @@ describe('CLI Test', () => {
     it('should show help msg when arg not passed to grep', async () => {
       const yargsCmd = yargs.command(cli)
       const output = await new Promise((resolve) => {
-        yargsCmd.parse('cli -g', (err, argv, output) => {
-          if (err) { }
-
+        yargsCmd.parse('cli -g', (_, argv, output) => {
           resolve(output)
         })
       })
@@ -119,23 +149,32 @@ describe('CLI Test', () => {
       expect(output).to.contain('baselineLimit')
       expect(output).to.contain('logLimit')
     })
+
+    it('should properly apply grep value to benchmarks', async () => {
+      const yargsCmd = yargs.command(cli)
+      const yargsResult = yargsCmd.parse(
+        'cli --grep console-baseline', {}
+      )
+
+      await start(benchmarks, yargsResult)
+    })
   })
 
-  describe('stressLimit', () => {
+  describe('stressLimit', function () {
+    this.timeout(11000)
+
     it('should parse stressLimit value', () => {
       const yargsCmd = yargs.command(cli)
       const yargsResult = yargsCmd.parse(
-        'cli --stressLimit 300', {}
+        'cli --stressLimit 10', {}
       )
-      expect(yargsResult.stressLimit).to.equal(300)
+      expect(yargsResult.stressLimit).to.equal(10)
     })
 
     it('should show help msg when arg not passed to stressLimit', async () => {
       const yargsCmd = yargs.command(cli)
       const output = await new Promise((resolve) => {
-        yargsCmd.parse('cli --stressLimit', (err, argv, output) => {
-          if (err) { }
-
+        yargsCmd.parse('cli --stressLimit', (_, argv, output) => {
           resolve(output)
         })
       })
@@ -147,9 +186,20 @@ describe('CLI Test', () => {
       expect(output).to.contain('baselineLimit')
       expect(output).to.contain('logLimit')
     })
+
+    it('should apply stressLimit arg', async () => {
+      const yargsCmd = yargs.command(cli)
+      const yargsResult = yargsCmd.parse(
+        'cli --stressLimit 10', {}
+      )
+
+      await start(benchmarks, yargsResult)
+    })
   })
 
-  describe('baselineLimit', () => {
+  describe('baselineLimit', function () {
+    this.timeout(6000)
+
     it('should parse baselineLimit value', () => {
       const yargsCmd = yargs.command(cli)
       const yargsResult = yargsCmd.parse(
@@ -161,9 +211,7 @@ describe('CLI Test', () => {
     it('should show help msg when arg not passed to baselineLimit', async () => {
       const yargsCmd = yargs.command(cli)
       const output = await new Promise((resolve) => {
-        yargsCmd.parse('cli --baselineLimit', (err, argv, output) => {
-          if (err) { }
-
+        yargsCmd.parse('cli --baselineLimit', (_, argv, output) => {
           resolve(output)
         })
       })
@@ -175,9 +223,20 @@ describe('CLI Test', () => {
       expect(output).to.contain('baselineLimit')
       expect(output).to.contain('logLimit')
     })
+
+    it('should apply baseLineLimit arg', async () => {
+      const yargsCmd = yargs.command(cli)
+      const yargsResult = yargsCmd.parse(
+        'cli --baselineLimit 1000', {}
+      )
+
+      await start(benchmarks, yargsResult)
+    })
   })
 
-  describe('logLimit', () => {
+  describe('logLimit', function () {
+    this.timeout(6000)
+
     it('should parse logLimit value', () => {
       const yargsCmd = yargs.command(cli)
       const yargsResult = yargsCmd.parse(
@@ -189,9 +248,7 @@ describe('CLI Test', () => {
     it('should show help msg when arg not passed to logLimit', async () => {
       const yargsCmd = yargs.command(cli)
       const output = await new Promise((resolve) => {
-        yargsCmd.parse('cli --logLimit', (err, argv, output) => {
-          if (err) { }
-
+        yargsCmd.parse('cli --logLimit', (_, argv, output) => {
           resolve(output)
         })
       })
@@ -202,6 +259,15 @@ describe('CLI Test', () => {
       expect(output).to.contain('stressLimit')
       expect(output).to.contain('baselineLimit')
       expect(output).to.contain('logLimit')
+    })
+
+    it('should apply logLimit arg', async () => {
+      const yargsCmd = yargs.command(cli)
+      const yargsResult = yargsCmd.parse(
+        'cli --logLimit 50', {}
+      )
+
+      await start(benchmarks, yargsResult)
     })
   })
 })
